requirements.txt
amqp==5.0.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-rest-auth==2.1.2
Django==3.1.4
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.6.0
django-recurrence==1.10.3
django-shortuuidfield==0.1.3
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-flex-fields==0.8.8
fcm-django==0.3.7
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.13
pyfcm==1.4.9
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
